orchestrator/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
langchain-core>=0.1.27
langchain>=0.0.350
langchain-community>=0.0.20
langsmith>=0.1.0
langchain-openai>=0.0.8
openai>=1.3.0
httpx>=0.26.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
sentence-transformers>=2.2.0
numpy>=1.21.0
faiss-cpu>=1.7.4 